requirements.txt
PyMuPDF>=1.23.0
python-docx>=1.1.0
keybert>=0.8.0
sentence-transformers>=2.7.0
scikit-learn>=1.4.0
nltk>=3.8.0
requests>=2.31.0
flask>=3.0.0
colorama>=0.4.6
watchdog>=4.0.0
pytest>=8.0.0
